Make it safer to extract revisions, even from example modules.
In some cases, modules do not have any revision.  These are mainly
example modules that don't follow strict formatting.

+++ tools/ietfYangDraftPull/draftPullLocal.py
@@ -16,6 +16,16 @@ from tools.utility.util import get_curr_dir
 
 LOGGER = log.get_logger('draftPullLocal')
 
+def get_latest_revision(f):
+    module = yangParser.parse(f)
+    if module.i_latest_revision is not None:
+        return module.i_latest_revision
+
+    rev = module.search_one('revision')
+    if rev is None:
+        return None
+
+    return rev.arg
 
 def load_json_from_url(url):
     failed = True
@@ -46,7 +56,9 @@ def check_name_no_revision_exist(directory):
                 exists = os.path.exists(directory + yang_file_name)
                 if exists:
                     parsed_yang = yangParser.parse(os.path.abspath(directory + yang_file_name))
-                    comapred_revision = parsed_yang.search('revision')[0].arg
+                    compared_revision = get_latest_revision(s.path.abspath(directory + yang_file_name))
+                    if compared_revision is None:
+                        continue
                     if revision == comapred_revision:
                         os.remove(directory + yang_file_name)
 
@@ -62,9 +74,9 @@ def check_early_revisions(directory):
                     files_to_delete.append(f2)
                     revision = f2.split(fname)[1].split('.')[0].replace('@', '')
                     if revision == '':
-                        revision = (yangParser.parse(os.path.abspath(directory
-                                                                     + f2))
-                                    .search('revision')[0].arg)
+                        revision = get_latest_revision(os.path.abspath(directory + f2))
+                        if revision is None:
+                            continue
                     year = int(revision.split('-')[0])
                     month = int(revision.split('-')[1])
                     day = int(revision.split('-')[2])
